refactor: LadderGenerator -> Ladder 역할 수정
