+++ src/main/java/ladder/LadderApplication.java
@@ -2,7 +2,6 @@ package ladder;
 
 import java.util.Scanner;
 import ladder.controller.LadderController;
-import ladder.domain.generator.LadderGenerator;
 import ladder.view.InputView;
 import ladder.view.OutputView;
 
@@ -11,9 +10,8 @@ public class LadderApplication {
     public static void main(String[] args) {
         final var inputView = new InputView(new Scanner(System.in));
         final var outputView = new OutputView();
-        final var ladderGenerator = new LadderGenerator();
 
-        LadderController ladderController = new LadderController(inputView, outputView, ladderGenerator);
+        LadderController ladderController = new LadderController(inputView, outputView);
         ladderController.run();
     }
 }
